+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## TBD
+
+### Bug fixes
+
+* Prevent ConcurrentModificationException thrown from Metadata class
+  [#935](https://github.com/bugsnag/bugsnag-android/pull/935)
+
 ## 5.1.0 (2020-09-08)
 
 ### Enhancements